Conflicts:
	model/houses.php

+++ controller/user.php
@@ -4,8 +4,9 @@ require('model/editionCompteTreatment.php');
 require('model/ajoutLogementTreatment.php');
 require('model/houses.php');
 
+
 if (!isset($_GET['page']) || empty($_GET['page'])) {
-    $page = "dashboard";
+    $page = "reception";
 } else {
     $page = $_GET['page'];
 }
@@ -14,9 +15,22 @@ $alerte = false;
 
 
 switch ($page) {
+    case 'reception' :
+        $tab = 'reception';
+        $title = 'reception';
+        $houses = getHouses($_SESSION['userID']);
+        break;
+
+
     case 'dashboard' :
         $tab = 'user-dashboard';
         $title = 'Dashboard';
+        $houses = getHouses($_SESSION['userID']);
+        break;
+
+    case 'dashboard-conso' :
+        $tab = 'user-dashboard-conso';
+        $title = 'Dashboard-Consommation';
         $houses = getHouses($_SESSION['userID']);
         break;
 
@@ -25,11 +39,26 @@ switch ($page) {
         $title = 'Mes logements';
         $houses = getHouses($_SESSION['userID']);
         if (isset($_POST['delete'])) {
-            if (isset($_POST['idHouse']))
-                deleteHouse($_POST['idHouse']);
-            else
+            $delete = true;
+            if (isset($_POST['idHouse'])) {
+                $rooms = getRooms($_POST['idHouse']);
+                foreach ($rooms as $room) {
+                    if (hasNoProduct($room['ID'])) {
+                        deleteRoom($room['ID']);
+                    } else {
+                        $alerte = "Impossible de supprimer une pièce qui contient des produits !";
+                        $delete = false;
+                    }
+                   break;
+                }
+                if ($delete) {
+                    deleteHouse($_POST['idHouse']);
+                    header("Refresh:0");
+                }
+            } else {
                 deleteProduct($_POST['idProduct']);
-            header("Refresh:0");
+                header("Refresh:0");
+            }
         }
         break;
 
@@ -121,10 +150,52 @@ switch ($page) {
             echo "<script type='text/javascript'>document.location.replace('index.php?target=user&page=logements');</script>";
             exit();
         }
+        break;
 
+    case 'edit-house' :
+        $tab = "edit-house";
+        $title = "Edition maison";
+        if (!isset($_GET['idhouse'])) {
+            echo "<script type='text/javascript'>document.location.replace('index.php?target=home&page=404');</script>";
+            exit();
+        }
+        $idHouse = $_GET['idhouse'];
+        if (isset($_POST['adresse']) AND !empty($_POST['adresse']) AND isset($_POST['nbrPieces']) AND !empty($_POST['nbrPieces']) AND isset($_POST['nbrHabitants']) AND !empty($_POST['nbrHabitants']) AND isset($_POST['superficie']) AND !empty($_POST['superficie'])) {
+            $adresse = htmlspecialchars_decode($_POST['adresse']);
+            $nbrPieces = htmlspecialchars($_POST['nbrPieces']);
+            $nbrHabitants = htmlspecialchars($_POST['nbrHabitants']);
+            $superficie = htmlspecialchars($_POST['superficie']);
+            updateLogements($adresse, $nbrHabitants, $nbrPieces, $superficie, $idHouse);
+            echo "<script type='text/javascript'>document.location.replace('index.php?target=user&page=logements');</script>";
+            exit();
+        }
+        if (isset($_POST['delete'])) {
+            if (isset($_POST['idRoom'])) {
+                if (hasNoProduct($_POST['idRoom'])) {
+                    deleteRoom($_POST['idRoom']);
+                    header("Refresh:0");
+                } else
+                    $alerte = "Impossible de supprimer une pièce qui contient des produits !";
 
+            }
+        }
+        break;
 
-
+    case 'edit-product' :
+        $tab = "user-edit-product";
+        $title = "Edition de produit";
+        if (!isset($_GET['idproduct'])) {
+            echo "<script type='text/javascript'>document.location.replace('index.php?target=home&page=404');</script>";
+            exit();
+        }
+        $idproduct = $_GET['idproduct'];
+        $productInfos = getProductInfos($idproduct);
+        $houses = getHouses($_SESSION['userID']);
+        if (isset($_POST['idPiece'])) {
+            moveProduct($idproduct, $_POST['idPiece']);
+            echo "<script type='text/javascript'>document.location.replace('index.php?target=user&page=logements');</script>";
+            exit();
+        }
         break;
 
     default :
